Simplified the running of tests wherever you are in the directory tree.
Bonus fix of getting ESLint to detect when we're using tsconfig paths to resolve imports.

Had to rename gql to graphql in neo-push as for some reason @neo4j/graphql was trying to resolve the graphql library from that directory.

Decided to export the types needed for OGM rather than importing them from dist, it was causing issues.